chore: docs, move storybook-generated files to the right place
