+++ frontend/.storybook/main.ts
@@ -2,8 +2,8 @@ import type { StorybookConfig } from '@storybook-vue/nuxt'
 
 const config: StorybookConfig = {
   stories: [
-    '../components/**/*.mdx',
-    '../components/**/*.stories.@(js|jsx|ts|tsx|mdx)',
+    '../app/components/**/*.mdx',
+    '../app/components/**/*.stories.@(js|jsx|ts|tsx|mdx)',
   ],
   addons: [
     '@storybook/addon-a11y',
