+++ frontend/app/components/assets/avif-test-image.avif
